fix #204 no render if selectedItem is not found
When using dynamically changing `items` props and there is a previously `state.selectedItem` set, if the new new `items` list coming in from props does not contain the same length and the `state.selectedItem` index in the Android native code does not exist, the `ArrayIndexOutOfBoundsException` error is thrown in the native code causing an app crash. This fix will return null if the `state.selectedItem` is not found in the newly rendered `items` list for the react-native Picker component.